isimler daha anlamlı hale getirilidi. fiziksel limitleri kontrol etmesi için bir aktör daha eklendi. test edilmeli! (closes #3)
