+++ cabin_simulator.py
@@ -1,6 +1,6 @@
 from aktos_dcs import *
 
-class Motor(Actor):
+class Cabin(Actor):
     # 3 yon mesaji var
     # her seferinde 1 cm ciksin
     # yukseklik degiskenini tutsun
@@ -8,20 +8,20 @@ class Motor(Actor):
         Actor.__init__(self)
         self.direction = "stop"
         self.cabin_height = 0
-        self.move_cabin()
+        Limiter()
 
-    def handle_MotorMessage(self,msg):
+    def handle_MotorMessage(self, msg):
         self.direction = msg.direction
 
-    def move_cabin(self):
+    def action(self):
         while True:
             if self.direction == "up":
-                self.cabin_height = self.cabin_height + 1
+                self.cabin_height += 1
 
                 self.send(ViewerMessage(height=self.cabin_height))
 
             elif self.direction == "down":
-                self.cabin_height = self.cabin_height - 1
+                self.cabin_height -= 1
                 self.send(ViewerMessage(height=self.cabin_height))
 
             sleep(1) # update interval
@@ -31,8 +31,20 @@ class Viewer(Actor):
         print time.time(), msg.height
 
 
+class Limiter(Actor):
+    def __init__(self):
+        Actor.__init__(self)
+        self.elevator_height = 20  # meters
+
+    def handle_ViewerMessage(self, msg):
+        if (msg.height > self.elevator_height or
+            msg.height < 0):
+            print "should not exceed physical limits of elevator. Stopping."
+            self.send(MotorMessage(direction="stop"))
+
+
 if __name__ == "__main__":
     ProxyActor()
-    view = Viewer()
-    motor = Motor()
-    joinall([view, motor])
+    Viewer()
+    Cabin()
+    wait_all()
